Limit to specific genes using bed file
This version of the code limits the analysis to a specific set of genes specified in a BED file, such as "TableI.genes.bed". Was used in the paper (S. L. O'Beirne et al., in review) to replicate genes associated with T2D in Qatar in the T2D-Genes cohort. Great for quick analysis of large datasets. PYTHON SKILLS NEEDED: If you want to modify the analysis, you can configure using the "full == 0" flags within the main script "ega.t2d.replication.narrow.py".